pyproject.toml
[tool.poetry]
name = "qwacker"
version = "0.1.0"
description = ""
authors = ["mosguinz <mos.guinz@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "^2.3.2"
async-timeout = "^4.0.3"
python-dotenv = "^1.0.1"
emoji = "^2.12.1"

[tool.poetry.dev-dependencies]
black = "^24.4.2"

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "bot.__main__:main"